pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zoho-creator-agent"
version = "0.1.0"
description = "Terminal AI agent for Zoho Creator with DuckDB + LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.1.0",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "requests>=2.32.0",
  "rich>=13.7.1",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[project.scripts]
agent = "agent.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
